[feat] : #6 HomeScreen 영화 리스트 + 제목 컴포넌트 추가
